chore: add normalize whitespace prism plugin
When passing code sample as children to a component, React add an extra
newline at the end of the sample. Both line-numbers and command-line
plugins use the newline character. So each code blocks was printing an
extra new line at the end. It is not wanted. The normalize whitespace
plugin prevent this.