fix(pages): do not show spinner when data are available
The `useSWR` hook set `isLoading` to `true` on first fetch even if
the fallback data is provided. It affects how pages are generated on
build: the loading page is showing instead of the actual page. So we
need to return a custom `isLoading` variable that checks if there is no
data.